modules/location/gRPC/requirements.txt
kafka-python==2.0.2
grpcio==1.47.0
protobuf==3.20.1
grpcio-tools==1.47.0